Support whitespace around variable names
With reference to the mustache spec at github.com/mustache/spec, the
current definition for tag content:

> The tag's content MUST be a non-whitespace character sequence NOT
> containing the current closing delimiter.

However, later tests state “Superfluous in-tag whitespace should be
ignored.” and include tests for `{{ string }}` and `{{{ string }}}`.

That would tend to suggest that whitespace is allowed between the
opening and closing delimiters, and the variable name. That’s how other
implementations seem to work.