Implement partials module
Partials require appropriate indentation, and also handling recursive
partials.

Added very basic environment as a type of symbol table for dealing with
recursive partials.

“Partial tags are used to expand an external template into the current
template.

The tag's content MUST be a non-whitespace character sequence NOT
containing the current closing delimiter.

This tag's content names the partial to inject.  Set Delimiter tags MUST
NOT affect the parsing of a partial.  The partial MUST be rendered
against the context stack local to the tag.  If the named partial cannot
be found, the empty string SHOULD be used instead, as in interpolations.

Partial tags SHOULD be treated as standalone when appropriate. If this
tag is used standalone, any whitespace preceding the tag should treated
as indentation, and prepended to each line of the partial before
rendering.”