Implement lambdas module
See https://github.com/mustache/spec/pull/85 for how the tests were
generated. If / when that pull request is merged, I will also update the
submodule reference in this project.

“Lambdas are a special-cased data type for use in interpolations and
sections.

“When used as the data value for an Interpolation tag, the lambda MUST
be treatable as an arity 0 function, and invoked as such.  The returned
value MUST be rendered against the default delimiters, then interpolated
in place of the lambda.

When used as the data value for a Section tag, the lambda MUST be
treatable as an arity 1 function, and invoked as such (passing a String containing the unprocessed section contents).  The returned value MUST
be rendered against the current delimiters, then interpolated in place
of the section.”